requirements.txt
tensorflow==2.2.0
scikit-image==0.16.2
protobuf==3.20.0
uvicorn
fastapi
Pillow
numpy
scipy